feat: clean up API
