+++ apps/next-app/pages/modules/[questionBank]/tests/index.page.tsx
@@ -1,6 +1,6 @@
 import * as fs from "node:fs/promises";
 import { AppHead } from "@chair-flight/react/components";
-import { LayoutModule, TestsOverview } from "@chair-flight/react/containers";
+import { LayoutModule, TestSearch } from "@chair-flight/react/containers";
 import { staticHandler } from "@chair-flight/trpc/server";
 import type { QuestionBankName } from "@chair-flight/base/types";
 import type { Breadcrumbs } from "@chair-flight/react/containers";
@@ -23,7 +23,7 @@ const Page: NextPage<PageProps> = ({ questionBank }) => {
   return (
     <LayoutModule fixedHeight questionBank={questionBank} breadcrumbs={crumbs}>
       <AppHead />
-      <TestsOverview questionBank={questionBank} sx={{ height: "100%" }} />
+      <TestSearch questionBank={questionBank} sx={{ height: "100%" }} />
     </LayoutModule>
   );
 };
@@ -31,7 +31,7 @@ const Page: NextPage<PageProps> = ({ questionBank }) => {
 export const getStaticProps = staticHandler<PageProps, PageParams>(
   async ({ params, helper }) => {
     await LayoutModule.getData({ helper, params });
-    await TestsOverview.getData({ helper, params });
+    await TestSearch.getData({ helper, params });
     return { props: params };
   },
   fs,
