+++ libs/core/app/src/index.ts
@@ -4,3 +4,4 @@ export * from "./questions/get-question-preview";
 export * from "./questions/get-new-variant";
 export * from "./random/random";
 export * from "./tests/create-test";
+export * from "./tests/process-test";
