+++ libs/react/containers/src/tests/test-search/index.ts
@@ -1,0 +1,1 @@
+export { TestSearch } from "./test-search";
