--- libs/react/containers/src/tests/tests-overview/index.ts
@@ -1,1 +1,0 @@
-export { TestsOverview } from "./tests-overview";
